chap3 working v2

+++ Chap3/ReactiveWebStore/app/models/Product.scala
@@ -4,7 +4,7 @@ case class Product
  ( var id:Option[Long],
    var name:String,
    var details:String,
-   var price:Long ) 
+   var price:BigDecimal ) 
 {
   override def toString:String = {
     "Product { id: " + id.getOrElse(0) + ",name: " + name + ", details: "+ details + ", price: " + price  + "}"
